Switched to a new app versioning scheme.
We are now going to number every version by year.minor.revision(build)

Every posted build will need increase the build number. There is no beta build number anymore.

The app will display the commit hash too and whether the local repo was clean or not during build which can be used later for debugging.

Also added a missing script that can be used to bump the version number on macOS (for now).